==> 07_PyCode/11_AN_DaskLocalJupyter_v1.ipynb <==
import os
import pandas as pd
import numpy as np

import pyreadr

#import dask_mpi as dm
from dask.distributed import Client, progress
import dask
from dask.distributed import performance_report

from datetime import datetime
from joblib import Parallel, delayed
from sklearn.ensemble import RandomForestRegressor
from joblib import dump, load
import joblib

import dask.config
import distributed

os.environ["MALLOC_TRIM_THRESHOLD_"] = str(dask.config.get("distributed.nanny.environ.MALLOC_TRIM_THRESHOLD_"))

!pwd

DP02_location = "/home/usr6/q70176a/DP02/"
DP02_result_location = "/home/usr6/q70176a/DP02/08_PyResults/"

DP02_location = "/Users/lichao/Library/CloudStorage/OneDrive-KyushuUniversity/02_Article/03_RStudio/"
DP02_result_location = "/Users/lichao/Library/CloudStorage/OneDrive-KyushuUniversity/02_Article/03_RStudio/08_PyResults/"

DP02_location = "D:/OneDrive - Kyushu University/02_Article/03_RStudio/"
DP02_result_location = "D:/OneDrive - Kyushu University/02_Article/03_RStudio/08_PyResults/"

client = Client(n_workers=3, nthreads=1)
#client = Client(n_workers=7, threads_per_worker=1)

client

dataset = pyreadr.read_r(DP02_location + "02_Data/SP_Data_49Variable_Weights_changeRangeOfLandCover_RdsVer.Rds")
dataset = dataset[None]
y = np.array(dataset.iloc[:, 0:1].values.flatten(), dtype='float64')
X = np.array(dataset.iloc[:, 1:50], dtype='float64')

%%time
model = RandomForestRegressor(n_estimators=1000, oob_score=True, 
                               random_state=1, max_features = 9, n_jobs=-1)
with joblib.parallel_backend("dask"): model.fit(X, y)

# SHAP
import dalex as dx
model_rf_exp = dx.Explainer(model, X, y, label = "RF Pipeline")

with performance_report(filename="dask-report-scatter.html"):
    X_scattered = client.scatter(X)
    model_rf_exp_scattered = client.scatter(model_rf_exp)

def singleSHAPprocess(obs_num, X, model_rf_exp):
    test_obs = X[obs_num:obs_num+1,:]
    shap_test = model_rf_exp.predict_parts(test_obs, type = 'shap', 
                                           B = 10, N = 900)
    result = shap_test.result[shap_test.result.B == 0]
    result = result[['contribution', 'variable_name']]
    result = result.transpose()
    result = result.rename(columns=result.iloc[1])
    result = result.drop(['variable_name'], axis=0)
    result = result.reset_index(drop=True)
    return result

%%time
delayed_results = [dask.delayed(singleSHAPprocess)(obs_num, X_scattered, model_rf_exp_scattered)  for obs_num in list(range(100))]
#results = client.compute(*delayed_results)

delayed_results[1]

with performance_report(filename="dask-report-shap.html"):
    %time results = dask.compute(*delayed_results)

client.close()

%%time
singleSHAPprocess(1, X, model_rf_exp)

import shap

explainer = shap.TreeExplainer(model)

choosen_instance = X[0]

choosen_instance

%%time
shap_values = explainer.shap_values(choosen_instance)

shap_values

%%time
#shap_values_ori=singleSHAPprocess(0, X, model_rf_exp) 

shap_values_ori

run = explainer.shap_values

%%time
run(choosen_instance)

client = Client(n_workers=2, nthreads=1)

client

run_scattered = client.scatter(run)

X_scattered = client.scatter(X)

def singleSHAP(obs_num, X, run):
    test_obs = X[obs_num:obs_num+1,:]
    result = run(test_obs)
    return result

%%time
delayed_results = [dask.delayed(singleSHAP)(obs_num, X_scattered, run_scattered)  for obs_num in list(range(1000))]
#results = client.compute(*delayed_results)

%%time
results = dask.compute(*delayed_results)

client.close()

## To HPC Test Memory

import os
import pandas as pd
import numpy as np

import pyreadr

#import dask_mpi as dm
from dask.distributed import Client, progress, get_task_stream
import dask

from datetime import datetime
from joblib import Parallel, delayed
from sklearn.ensemble import RandomForestRegressor
from joblib import dump, load
import joblib

DP02_location = "/home/usr6/q70176a/DP02/"
DP02_result_location = "/home/usr6/q70176a/DP02/08_PyResults/"

client = Client(n_workers=3, threads_per_worker=1)
print(client)

dataset = pyreadr.read_r(DP02_location + "02_Data/SP_Data_49Variable_Weights_changeRangeOfLandCover_RdsVer.Rds")
dataset = dataset[None]
y = np.array(dataset.iloc[:, 0:1].values.flatten(), dtype='float64')
X = np.array(dataset.iloc[:, 1:50], dtype='float64')


model = RandomForestRegressor(n_estimators=1000, oob_score=True, 
                               random_state=1, max_features = 9, n_jobs=-1)
from dask.distributed import performance_report
with performance_report(filename="dask-report.html"):
    with joblib.parallel_backend("dask"): model.fit(X, y)

import shap
explainer = shap.TreeExplainer(model)
def singleSHAP(obs_num, X, explainer):
    test_obs = X[obs_num:obs_num+1,:]
    result = explainer.shap_values(test_obs)
    return result


with performance_report(filename="dask-report-shap.html"):
    delayed_results = [dask.delayed(singleSHAP)(obs_num, X, explainer)  for obs_num in list(range(20))]
    results = dask.compute(*delayed_results)

with performance_report(filename="dask-report-scatter.html"):
    X_scattered = client.scatter(X)
    explainer_scattered = client.scatter(explainer)

test = client.submit(singleSHAP, 0, X_scattered, run_scattered)

seriesTest = []
for obs in list(range(20)):
    test = client.submit(singleSHAP, obs, X_scattered, run_scattered)
    seriesTest.append(test)

client.gather(seriesTest)

client.close()

with performance_report(filename="dask-report-shap.html"):
    delayed_results = [dask.delayed(singleSHAP)(obs_num, X_scattered, run_scattered)  for obs_num in list(range(20))]
    results = dask.compute(*delayed_results)
